chandata: cast int to uint16 instead opposite in IsChannelData
Before:
BenchmarkIsChannelData-12    	2000000000	         1.73 ns/op	6350.93 MB/s	       0 B/op	       0 allocs/op

After:
BenchmarkIsChannelData-12    	2000000000	         1.64 ns/op	6698.72 MB/s	       0 B/op	       0 allocs/op

+++ chandata.go
@@ -115,5 +115,5 @@ func IsChannelData(buf []byte) bool {
 	}
 	// Check that length is valid.
 	l := bin.Uint16(buf[channelNumberSize:channelDataHeaderSize])
-	return int(l) == len(buf[channelDataHeaderSize:])
+	return l == uint16(len(buf[channelDataHeaderSize:]))
 }
